Removing mention of Docker Machine
Docker Machine is not supported because it does not support file change
notifications. File change notifications are not sent to the VM which in
turn doesn't trigger a new build. While there are hacks around it, I'm
going to just remove support for Docker Machine instead since it should
be deprecated soon.